requirements.txt
asgiref==3.2.10
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
boto3==1.21.18
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
Collectfast==2.2.0
colorama==0.4.4
cryptography==36.0.1
defusedxml==0.7.1
Django==3.1
django-allauth==0.49.0
django-autoslug==1.9.8
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-currentuser==0.5.3
django-js-asset==2.0.0
django-modeltranslation==0.17.5
django-mptt==0.13.4
django-storages==1.12.3
django-translations==1.3.0
djangorestframework==3.13.1
gunicorn==20.1.0
heroku3==5.1.4
idna==3.3
jmespath==0.10.0
joblib==1.1.0
nltk==3.7
numpy==1.22.3
oauthlib==3.2.0
pandas==1.4.1
Pillow==9.0.1
polib==1.1.1
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
regex==2022.3.2
requests==2.27.1
requests-oauthlib==1.3.1
rosetta==0.3
s3transfer==0.5.1
scipy==1.8.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tqdm==4.63.0
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8
whitenoise==6.0.0
botocore~=1.24.46
setuptools~=60.2.0